move blending-related nodes and code to `mod blending_nodes`
