+++ editor/src/messages/portfolio/document_migration.rs
@@ -31,6 +31,9 @@ const REPLACEMENTS: &[(&str, &str)] = &[
 	("graphene_core::ops::Vector2ValueNode", "graphene_core::ops::CoordinateValueNode"),
 	("graphene_core::raster::BlackAndWhiteNode", "graphene_core::raster::adjustments::BlackAndWhiteNode"),
 	("graphene_core::raster::BlendNode", "graphene_core::raster::adjustments::BlendNode"),
+	("graphene_core::raster::BlendModeNode", "graphene_core::blending_nodes::BlendModeNode"),
+	("graphene_core::raster::OpacityNode", "graphene_core::blending_nodes::OpacityNode"),
+	("graphene_core::raster::BlendingNode", "graphene_core::blending_nodes::BlendingNode"),
 	("graphene_core::raster::ChannelMixerNode", "graphene_core::raster::adjustments::ChannelMixerNode"),
 	("graphene_core::raster::adjustments::ColorOverlayNode", "graphene_core::raster::adjustments::ColorOverlayNode"),
 	("graphene_core::raster::ExposureNode", "graphene_core::raster::adjustments::ExposureNode"),
